Combine libs dev, header, input, and assignment funcs
